DarkStarScoringSystem/requirements.txt
firebase-admin>=6.0.0
playwright>=1.40.0
requests>=2.31.0
jsonschema>=4.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0

